Add the whole US
Adds the whole of US as a "state". US has different population and history data files, but the output is treated as it were a state. Made changes to mortality tables to replace infected - 14 mortality rate with the current daily mortality rate as a percentage of the current number of daily new cases.