Fixed smoothing of % positive
Instead of smoothing % positive, compute from smoothed tested and positives. Due to bad data, sometimes new positives are reported, but negatives are not, leading to 100% positives, which corrupts the smoothed curve. Also fixed hardcoded 7 in smoothing function, changed tested support to 7 (from 21) and a few other small fixes and improvements.